Restore xhtml5 but rename output files
The problem with xhtml and firefox is due to me naming the output
files as ".html".  It seems Chrome works around being sent XHTML with
the wrong mimetype, but Firefox doesn't.

The solution is to keep using the XHTML5 docbook output, but rename
the files to .xhtml so apache sends it with the right mimetype.  This
way firefox picks it up correctly and uses the right character
encoding.  To avoid breakage, I will do a redirect on the server side
from html->xhtml

Closes #13